rename infowigets to dataEntryOutput
