+++ public/manifest.webapp
@@ -19,5 +19,5 @@
     "name": "HISP",
     "url": "http://dhis2.org"
   },
-  "manifest_generated_at": "Tue Aug 14 2018 09:20:35 GMT+0200 (CEST)"
+  "manifest_generated_at": "Wed Aug 15 2018 13:50:33 GMT+0200 (CEST)"
 }
