+++ src/core_modules/capture-core/src/components/Pages/EditEvent/DataEntry/EditEventDataEntry.component.js
@@ -19,10 +19,10 @@ import withDefaultShouldUpdateInterface from
     '../../../../components/DataEntry/dataEntryField/withDefaultShouldUpdateInterface';
 import inMemoryFileStore from '../../../DataEntry/file/inMemoryFileStore';
 import withNotes from '../../../DataEntry/withNotes';
-import withIndicatorWidget from '../../../DataEntry/dataEntryInfoWidgets/withIndicatorWidget';
-import withFeedbackWidget from '../../../DataEntry/dataEntryInfoWidgets/withFeedbackWidget';
-import withErrorWidget from '../../../DataEntry/dataEntryInfoWidgets/withErrorWidget';
-import withWarningWidget from '../../../DataEntry/dataEntryInfoWidgets/withWarningWidget';
+import withIndicatorOutput from '../../../DataEntry/dataEntryOutput/withIndicatorOutput';
+import withFeedbackOutput from '../../../DataEntry/dataEntryOutput/withFeedbackOutput';
+import withErrorOutput from '../../../DataEntry/dataEntryOutput/withErrorOutput';
+import withWarningOutput from '../../../DataEntry/dataEntryOutput/withWarningOutput';
 
 const getStyles = () => ({
 });
@@ -105,11 +105,11 @@ const buildCompleteFieldSettingsFn = () => {
 
 const ReportDateField = withDataEntryField(buildReportDateSettingsFn())(DataEntry);
 const CompleteField = withDataEntryField(buildCompleteFieldSettingsFn())(ReportDateField);
-const FeedbackWidget = withFeedbackWidget()(CompleteField);
-const IndicatorWidget = withIndicatorWidget()(FeedbackWidget);
-const WarningWidget = withWarningWidget()(IndicatorWidget);
-const ErrorWidget = withErrorWidget()(WarningWidget);
-const SaveableDataEntry = withSaveButton(getSaveOptions)(ErrorWidget);
+const FeedbackOutput = withFeedbackOutput()(CompleteField);
+const IndicatorOutput = withIndicatorOutput()(FeedbackOutput);
+const WarningOutput = withWarningOutput()(IndicatorOutput);
+const ErrorOutput = withErrorOutput()(WarningOutput);
+const SaveableDataEntry = withSaveButton(getSaveOptions)(ErrorOutput);
 const NotesDataEntry = withNotes(buildNoteFieldSettingsFn)(SaveableDataEntry);
 const CancelableDataEntry = withCancelButton(getCancelOptions)(NotesDataEntry);
